+++ src/view/waypoint-item-view.js
