+++ src/view/waypoint-list-view.js
